+++ admin/dv_ldv/dv_act/dv_act.php
@@ -74,8 +74,8 @@
             <tbody>
               <td colspan="11">
                 <a href="?act=ct_dv_act&id=<?= $result['id_pk_dv'] ?>">Chi tiết</a>
-                <a href="?act=delete_dv_act&id=<?= $result['id_pk_dv'] ?>">Xóa</a>
-                <a href="?act=edit_act&id=<?= $result['id_dv_user']?>" class="ml-5">Cập Nhật Trạng Thái</a>
+                <a href="?act=delete_dv_act&id=<?= $result['id_pk_dv'] ?>" onclick="return confirm('Bạn chắn chắn muốn xóa?')">Xóa</a>
+                <a href="?act=edit_act&id=<?= $result['id_dv_user']?>" class="ml-5" onclick="return confirm('Bạn chắn chắn muốn cập nhật?')">Cập Nhật Trạng Thái</a>
               </td>
             </tbody>
             </tbody>
